SCREAM: completely clean up IC section for constant fields in input yaml files
The grid name list is no longer used. Notice that we should not have two copies
of the same field on two grids that are both input fields to the atm.
This makes no sense. Also, it is highly unlikely that a field not on the ref grid
should be an IC field (restarts are a different issue, but restarts do not have fields
inited to a constant anyways).